Refactor getArmy
