+++ src/components/input/army_builder.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { CardMultiSelect, CardSingleSelect } from 'components/info/card'
 import { getArmyBuilderCards } from './army_builder_cards'
 import { withSelectOne, withSelectMultiple } from 'utils/withSelect'
-import { getArmy } from 'utils/getArmy'
+import { getArmy } from 'utils/getArmy/getArmy'
 import { componentWithSize } from 'utils/mapSizesToProps'
 import { realmscape, selections, factionNames, army } from 'ducks'
 import { TSupportedFaction } from 'meta/factions'
