+++ src/components/input/savedArmies/load_army_btn.tsx
@@ -6,7 +6,7 @@ import { IArmy, TUnits } from 'types/army'
 import { ISavedArmyFromApi } from 'types/savedArmy'
 import { ISelections } from 'types/selections'
 import { TAllySelectionStore } from 'types/store'
-import { getArmy } from 'utils/getArmy'
+import { getArmy } from 'utils/getArmy/getArmy'
 import { logEvent } from 'utils/analytics'
 
 interface ILoadButtonProps {
