+++ src/components/input/toolbar.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
 import { without } from 'lodash'
 import is from 'is_js'
-import { getArmy } from 'utils/getArmy'
+import { getArmy } from 'utils/getArmy/getArmy'
 import { logPrintEvent, logClick } from 'utils/analytics'
 import { useSubscription } from 'context/useSubscription'
 import { factionNames, selections, army } from 'ducks'
